requirements.txt
Django>=3.2,<3.3
django-cleanup==6.0.0
djangorestframework>=3.13.1,<3.14.0
flake8>=4.0.1,<4.1.0
psycopg2-binary==2.9.3
pillow==9.0.1